Allow value comparison for booleans as well.
e.g. attr:a = false

This is needed so that we can differentiate between missing value and
false value when a filter is `not attr:a`. Before this change, this
filter would return true in both cases. After this change, we can use
`attr:a = false` to ensure it only returns true when attr:a is defined
and has false as value.